// api-public/src/nexus/generated/typings.ts
export interface NexusGenObjects {
// ...
  Game: {
    // root type
    competition?: string | null; // String
    coverImage?: string | null; // String
    date: number; // Float!
    gameID: number; // Int!
    headline?: string | null; // String
    highPoints?: number | null; // Int
    isReminded: boolean; // Boolean!
    leagueCode?: string | null; // String
    links?: Array<NexusGenRootTypes['Link'] | null> | null; // [Link]
    location?: string | null; // String
    points: number; // Int!
    pointsLevel?: string | null; // String
    sport: string; // String!
    team1City?: string | null; // String
    team1Color?: string | null; // String
    team1DisplayName?: string | null; // String
    team1ID?: number | null; // Int
    team1Name?: string | null; // String
    team1Nickname?: string | null; // String
    team1Ranking?: number | null; // Int
    team1Score?: number | null; // Int
    team2City?: string | null; // String
    team2Color?: string | null; // String
    team2DisplayName?: string | null; // String
    team2ID?: number | null; // Int
    team2Name?: string | null; // String
    team2Nickname?: string | null; // String
    team2Ranking?: number | null; // Int
    team2Score?: number | null; // Int
    timeLeft?: string | null; // String
  };
// ...
}

export interface NexusGenInterfaces {}

export interface NexusGenUnions {}

export type NexusGenRootTypes = NexusGenObjects;